Work around Firefox bug
I haven't been able to diagnose why this happens, but essentially I boiled down a bug I was facing to the fact that the default value for the exists parameter wasn't being used. Instead, exists would be a class reference at runtime. This is obviously a bug, and it only happened in Firefox. But after failing to reproduce it in a smaller example, I opted by just getting the argument from the arguments array, which seems to work.